+++ platformio/__main__.py
@@ -26,19 +26,35 @@ from platformio.compat import CYGWIN
 @click.command(cls=PlatformioCLI,
                context_settings=dict(help_option_names=["-h", "--help"]))
 @click.version_option(__version__, prog_name="PlatformIO")
-@click.option("--force",
-              "-f",
+@click.option("--force", "-f", is_flag=True, help="DEPRECATE")
+@click.option("--caller", "-c", help="Caller ID (service)")
+@click.option("--no-ansi",
               is_flag=True,
-              help="Force to accept any confirmation prompts.")
-@click.option("--caller", "-c", help="Caller ID (service).")
+              help="Do not print ANSI control characters")
 @click.pass_context
-def cli(ctx, force, caller):
+def cli(ctx, force, caller, no_ansi):
+    try:
+        if (no_ansi or str(
+                os.getenv(
+                    "PLATFORMIO_NO_ANSI",
+                    os.getenv("PLATFORMIO_DISABLE_COLOR"))).lower() == "true"):
+            # pylint: disable=protected-access
+            click._compat.isatty = lambda stream: False
+        elif str(
+                os.getenv(
+                    "PLATFORMIO_FORCE_ANSI",
+                    os.getenv("PLATFORMIO_FORCE_COLOR"))).lower() == "true":
+            # pylint: disable=protected-access
+            click._compat.isatty = lambda stream: True
+    except:  # pylint: disable=bare-except
+        pass
+
     maintenance.on_platformio_start(ctx, force, caller)
 
 
 @cli.resultcallback()
 @click.pass_context
-def process_result(ctx, result, force, caller):  # pylint: disable=W0613
+def process_result(ctx, result, *_, **__):
     maintenance.on_platformio_end(ctx, result)
 
 
@@ -53,16 +69,6 @@ def configure():
         import urllib3
         urllib3.disable_warnings()
     except (AttributeError, ImportError):
-        pass
-
-    try:
-        if str(os.getenv("PLATFORMIO_DISABLE_COLOR", "")).lower() == "true":
-            # pylint: disable=protected-access
-            click._compat.isatty = lambda stream: False
-        elif str(os.getenv("PLATFORMIO_FORCE_COLOR", "")).lower() == "true":
-            # pylint: disable=protected-access
-            click._compat.isatty = lambda stream: True
-    except:  # pylint: disable=bare-except
         pass
 
     # Handle IOError issue with VSCode's Terminal (Windows)
@@ -87,7 +93,7 @@ def main(argv=None):
         sys.argv = argv
     try:
         configure()
-        cli(None, None, None)
+        cli()  # pylint: disable=no-value-for-parameter
     except SystemExit:
         pass
     except Exception as e:  # pylint: disable=broad-except